fixed queries with datetime on csv data source, simplified service initialization
