+++ Earthquake.API.Tests/EarthquakeControllerWithCsvDataTest.cs
@@ -35,8 +35,8 @@ namespace Earthquake.API.Tests
         {
             const double latitude = 38.5563316;
             const double longitude = -119.5276642;
-            var startDate = DateTime.Parse("2021-10-12");
-            var endDate = DateTime.Parse("2021-10-13");
+            var startDate = DateTime.Parse("2021-10-18T00:41:49.039Z");
+            var endDate = DateTime.Parse("2021-10-18T00:43:16.039Z");
 
             var data = _controller.FindByCoordinatesBetweenDateRange(latitude, longitude, startDate, endDate);
 
@@ -47,10 +47,10 @@ namespace Earthquake.API.Tests
         [TestMethod]
         public void EndPoint_Returns404_Is_True()
         {
-            const double latitude = 90;
-            const double longitude = 90;
-            var startDate = DateTime.Parse("2024-01-01");
-            var endDate = DateTime.Parse("2025-01-01");
+            const double latitude = 38.5563316;
+            const double longitude = -119.5276642;
+            var startDate = DateTime.Parse("2021-10-18T00:27:33.842Z");
+            var endDate = DateTime.Parse("2021-10-18T00:41:49.039Z");
 
             var data = _controller.FindByCoordinatesBetweenDateRange(latitude, longitude, startDate, endDate);
