+++ Earthquake.API/Program.cs
@@ -1,11 +1,4 @@
-using System;
-using System.IO;
-using System.Reflection;
-using Earthquake.API.AutoMappingProfiles;
-using Earthquake.Data.CSV;
-using Earthquake.Data.USGS;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 
 namespace Earthquake.API
@@ -20,23 +13,7 @@ namespace Earthquake.API
         public static IHostBuilder CreateHostBuilder(string[] args)
         {
             return Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder => { webBuilder.UseStartup<Startup>(); })
-                .ConfigureServices((hostContext, services) =>
-                {
-                    var configuration = hostContext.Configuration;
-
-                    var usgsApiBaseAddress = configuration["USGSApiBaseAddress"];
-                    services.AddSingleton<IUsgsHttpClient>(new UsgsHttpClient(usgsApiBaseAddress));
-                    services.AddSingleton<IUsgsDataContext, UsgsDataContext>();
-
-                    var earthquakeDataCsvFileName = configuration["EarthquakeDataCSVFileName"];
-                    var csvFileDirectory = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
-                    var csvFullFileName = Path.Combine(csvFileDirectory ?? throw new InvalidOperationException(), earthquakeDataCsvFileName);
-                    services.AddSingleton<ICsvParser>(new CsvParser(csvFullFileName));
-                    services.AddSingleton<ICsvDataContext, CsvDataContext>();
-
-                    services.AddAutoMapper(typeof(EarthquakeMappingProfile));
-                });
+                .ConfigureWebHostDefaults(webBuilder => { webBuilder.UseStartup<Startup>(); });
         }
     }
 }
